Add files via upload
This program facilitates easiser funding of the ZenCash secure node multisg. Since
we can't send many utxos to the same address with z_sendmany, PART1 of this two-part
script sends precise amounts to a set of own addresses that will then be sent to 
the sec node multisig in PART2.   

Input files are a .txt with utxo amounts required and another .txt file with available
own wallet addresses.